Implement client-side recaptcha
There is a situation where it takes a lot of time for the recaptcha to be loaded,
which can result in a failed request to submit the email. Have to look into it.

Also, there is a weird console warning if using process.env for the client-side sitekey.

This commit makes the contact feature not work without javascript enabled.