Fix missing fields check not working
Because we added a onSubmit() to the contact form, the "required" html stopped working,
to fix it, this commit adds a javascript check. Also, it now checks on server-side too.

This commit also hard-codes the public recaptcha site key in the contact page. This was
necessary to fix a bug where sometimes the page would not recognize the key in process.env
and fail with an error. Don't know why that happens, but it is working now. Remember to
change the key when in production. It is ok to hard-code it, it's suposed to be public.

Also, removed some code from commit 3b28b27 that was not necessary anymore.